domain-order - OrderEntity 추가 - 주문시 User가 소유한 Product의 상태를 Ordered로 처리하기 위해서 ProductState 추가 - 요청시 맵핑될 VO인 OrderRequest 추가 - 오전 10시 ~ 오후 8시 사이에만 주문 가능하도록 도메인 규칙 구현 -

+++ src/main/java/com/photosend/photosendserver01/domains/user/presentation/UserProductController.java
@@ -29,7 +29,7 @@ public class UserProductController {
     }
 
     @DeleteMapping(value = "/{users-id}/product/{product-id}")
-    public void deleteProduct(@PathVariable("users-id") String userId, @PathVariable("product-id") Long productId) {
+    public void deleteProduct(@PathVariable("users-id") String userId, @PathVariable("product-id") Integer productId) {
         userProductService.deleteProductImage(userId, productId);
     }
